miatoll: wifi: Explicitly disable protected management frames
This fixes support with WPA2/3 mixed access points; the primary issue
was that while this device doesn't support PMF, said access points try
to query for PMF in order to check for WPA3 support, which causes our
WLAN driver to die a painful death.

Disable it from the overlay, as AOSP enables it by default.

It's a wonder why it was enabled in the first place.

Change-Id: I89dbce7ae4c91cea9f1828e8b627a828e2b08011
Signed-off-by: Arindam Bhattacharjee <abhattacharjee717@gmail.com>